1_project/notifier/requirements.txt
celery[redis]==5.2.7
flask==2.2.2
Pillow==7.2.0
structlog==22.3.0
pydantic==1.10.2
python-dotenv==0.21.0
